requirements.txt
firebase-admin==6.5.0
Flask==3.0.3
google-ai-generativelanguage==0.6.9